Refactor mplfinance candle plot
