+++ gamestonk_terminal/cryptocurrency/due_diligence/binance_model.py
@@ -4,18 +4,14 @@ __docformat__ = "numpy"
 import argparse
 import logging
 from collections import defaultdict
-from typing import List, Tuple, Union, Optional
+from typing import List, Tuple, Union
 
-import matplotlib.pyplot as plt
-import mplfinance as mpf
 import pandas as pd
 from binance.client import Client
 from binance.exceptions import BinanceAPIException
 
-from gamestonk_terminal.config_terminal import theme
 import gamestonk_terminal.config_terminal as cfg
 from gamestonk_terminal.decorators import log_start_end
-from gamestonk_terminal.helper_funcs import plot_autoscale
 from gamestonk_terminal.rich_config import console
 
 logger = logging.getLogger(__name__)
@@ -134,58 +130,58 @@ def show_available_pairs_for_given_symbol(
     return None, []
 
 
-@log_start_end(log=logger)
-def plot_candles(
-    candles_df: pd.DataFrame,
-    title: str,
-    external_axes: Optional[List[plt.Axes]] = None,
-) -> None:
-    """Plot candle chart from dataframe. [Source: Binance]
+# @log_start_end(log=logger)
+# def plot_candles(
+#     candles_df: pd.DataFrame,
+#     title: str,
+#     external_axes: Optional[List[plt.Axes]] = None,
+# ) -> None:
+#     """Plot candle chart from dataframe. [Source: Binance]
 
-    Parameters
-    ----------
-    candles_df: pd.DataFrame
-        Dataframe containing time and OHLCV
-    title: str
-        title of graph
-    external_axes : Optional[List[plt.Axes]], optional
-        External axes (1 axis is expected in the list), by default None
-    """
+#     Parameters
+#     ----------
+#     candles_df: pd.DataFrame
+#         Dataframe containing time and OHLCV
+#     title: str
+#         title of graph
+#     external_axes : Optional[List[plt.Axes]], optional
+#         External axes (1 axis is expected in the list), by default None
+#     """
 
-    candle_chart_kwargs = {
-        "type": "candle",
-        "style": theme.mpf_style,
-        "volume": True,
-        "xrotation": theme.xticks_rotation,
-        "scale_padding": {"left": 0.3, "right": 1, "top": 0.8, "bottom": 0.8},
-        "update_width_config": {
-            "candle_linewidth": 0.6,
-            "candle_width": 0.8,
-            "volume_linewidth": 0.8,
-            "volume_width": 0.8,
-        },
-        "warn_too_much_data": 10000,
-    }
+#     candle_chart_kwargs = {
+#         "type": "candle",
+#         "style": theme.mpf_style,
+#         "volume": True,
+#         "xrotation": theme.xticks_rotation,
+#         "scale_padding": {"left": 0.3, "right": 1, "top": 0.8, "bottom": 0.8},
+#         "update_width_config": {
+#             "candle_linewidth": 0.6,
+#             "candle_width": 0.8,
+#             "volume_linewidth": 0.8,
+#             "volume_width": 0.8,
+#         },
+#         "warn_too_much_data": 10000,
+#     }
 
-    # This plot has 2 axes
-    if not external_axes:
-        candle_chart_kwargs["returnfig"] = True
-        candle_chart_kwargs["figratio"] = (10, 7)
-        candle_chart_kwargs["figscale"] = 1.10
-        candle_chart_kwargs["figsize"] = plot_autoscale()
-        fig, _ = mpf.plot(candles_df, **candle_chart_kwargs)
-        fig.suptitle(
-            f"\n{title}",
-            x=0.055,
-            y=0.965,
-            horizontalalignment="left",
-        )
-        theme.visualize_output()
-    else:
-        if len(external_axes) != 2:
-            console.print("[red]Expected list of 2 axis items./n[/red]")
-            return
-        (ax, volume) = external_axes
-        candle_chart_kwargs["ax"] = ax
-        candle_chart_kwargs["volume"] = volume
-        mpf.plot(candles_df, **candle_chart_kwargs)
+#     # This plot has 2 axes
+#     if not external_axes:
+#         candle_chart_kwargs["returnfig"] = True
+#         candle_chart_kwargs["figratio"] = (10, 7)
+#         candle_chart_kwargs["figscale"] = 1.10
+#         candle_chart_kwargs["figsize"] = plot_autoscale()
+#         fig, _ = mpf.plot(candles_df, **candle_chart_kwargs)
+#         fig.suptitle(
+#             f"\n{title}",
+#             x=0.055,
+#             y=0.965,
+#             horizontalalignment="left",
+#         )
+#         theme.visualize_output()
+#     else:
+#         if len(external_axes) != 2:
+#             console.print("[red]Expected list of 2 axis items./n[/red]")
+#             return
+#         (ax, volume) = external_axes
+#         candle_chart_kwargs["ax"] = ax
+#         candle_chart_kwargs["volume"] = volume
+#         mpf.plot(candles_df, **candle_chart_kwargs)
