+++ gamestonk_terminal/helper_classes.py
@@ -277,14 +277,18 @@ class TerminalStyle:
         )
 
     # pylint: disable=import-outside-toplevel
-    def visualize_output(self):
+    def visualize_output(self, force_tight_layout: bool = True):
         """Show chart in an interactive widget."""
         import gamestonk_terminal.feature_flags as gtff
         from gamestonk_terminal.rich_config import console
+        import warnings
+
+        warnings.filterwarnings("ignore")
 
         if gtff.USE_WATERMARK:
             self.add_label(plt.gcf())
-        plt.tight_layout(pad=self.tight_layout_padding)
+        if force_tight_layout:
+            plt.tight_layout(pad=self.tight_layout_padding)
         if gtff.USE_ION:
             plt.ion()
         plt.show()
